Fixed USB OTG initialization bug.
The CrazyRadio PA dongle was failing to initialize on Android on some phones with a specific USB OTG cable. It is unclear exactly what was happening, but my suspicion is that the dongle firmware was failing to initialize USB immediately after power on, and then hanging. To address this, I added a simple timeout to `init_usb()` which attempts to re-initialize the USB stack if it is not successful on the first attempt. This appears to have resolved the problem.