Added support for non-Nordic transceivers.
The nRF24LU1+ dongles are useful for generic (non-Nordic) 2.4GHz transceivers. In order to accommodate this, I added functions `enter_promiscuous_mode_generic` and `transmit_payload_generic` to `nrf24`, and made corresponding changes to the firmware. This makes it possible to to communicate with other types of narrowband 2.4GHz GFSK transceivers.