Experimental Logitech dongle flashing support
Added experimental support to flash this firmware onto nRF24LU1+ based Logitech Unifying dongles, which is performed via `sudo make logitech_install`. This is currently a one-way operation, and the firmware does not support entering the Logitech bootloader again to flash back to the Logitech firmware, or update the research firmware.